Simplify implementation of off-thread
The original implementation used channel-put-event to avoid an obscure attack whereby the
code in the sandbox could hold up the worker thread by exiting before reading the response
from the channel.

This diff switches the implementation to buffered async channels which don't block on put
(up to the buffer size limit of 1).  Makes the implementation quite a bit easier to read.

Tested locally.